create transforms.go, case insensitive parse
